Simplify and generalize http-url regexp.
It's still strict about the protocol and host part so that we stil
have high confidence of not finding many false positives, but more
relaxed about the rest, so that we don't cut off too early.

The rest can now consist of any sequence of allowed characters, but
can't end in "," or ".".