[libcontentaction] Remove tracker related functionality. JB#54405 OMP#JOLLA-160
To my knowledge we haven't ever used any of these parts and Tracker3
is about to break anything that might still work, first due to d-bus
api version bump but also because there is no more such a central
tracker storage, instead the miner has mined information and the apps
can have their own databases.

In theory we could possibly create a new d-bus service implementing
the tracker api for fetching this kind of things, and then
reimplementing this with named graphs from named services, but
not seeing a point in that currently.